Created a new script for getting the highest occupation value
I decided to make a new script so i wouldn't have to have to keep getting requests on the bls website. I am trying to save each state with the occupation and rescpective value. It has not been fully adding to the dictionary correctly. Let me know if you have any suggestions.